refactor: Phase 3 - Extract service layer from main.py
- Create services module with ArxivService, OpenAIService, and integration services
- Extract arXiv paper retrieval to ArxivService
- Extract OpenAI filtering and summarization to OpenAIService
- Extract Discord, Google Drive, and Zotero operations to integration services
- Extract utility functions to services/utils.py
- Update main.py to use service layer with dependency injection pattern
- Maintain all functionality while improving code organization and testability
- All type checks and linting pass
- Test mode confirms refactoring works correctly
