+++ src/autojournalsummarizer/services/__init__.py
@@ -1,0 +1,17 @@
+"""Service layer modules for AutoJournalSummarizer."""
+
+from .arxiv import ArxivService
+from .integrations import DiscordService, GoogleDriveService, ZoteroService
+from .openai_service import OpenAIService
+from .utils import extract_text_from_pdf, get_last_published_datetime, update_log
+
+__all__ = [
+    "ArxivService",
+    "OpenAIService",
+    "DiscordService",
+    "GoogleDriveService",
+    "ZoteroService",
+    "extract_text_from_pdf",
+    "get_last_published_datetime",
+    "update_log",
+]
